Waypoint Updater node ready for test
Implemented
-- Partial waypoint updater
-- Full waypoint updater
It takes /traffic_waypoint and publishes /final_waypoints for the car to follow. Note that /traffic_waypoint needs to be published by the perception node. There could be a way the perception node can be made to publish the ground truth information from /vehicle/traffic_lights to test the full pipeline from planning to control.